pyproject.toml
[project]
name = "pyab1"
version = "0.1.0"
description = "Read Sanger sequencing ab1 files."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "matplotlib>=3.10.6",
    "polars>=1.0.0",
    "xlsxwriter>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyab1"]

[dependency-groups]
dev = [
    "basedpyright>=1.37.0",
    "ruff>=0.14.10",
]

[project.scripts]
process = "pyab1.process:main"
process-many = "pyab1.process_many:main"
